# docs/swagger.yaml
basePath: /api/v2
definitions:
  apperrors.ErrorResponse:
    properties:
      error_response:
        type: string
    type: object
...
  doc_structs.ListResponse:
    properties:
      list:
        $ref: '#/definitions/entities.List'
    type: object
  doc_structs.TaskResponse:
    properties:
      task:
        $ref: '#/definitions/entities.Task'
    type: object
...
  dto.ListID:
    properties:
      id:
        type: integer
    type: object
...
  dto.NewListInfo:
    properties:
      board_id:
        type: integer
      description:
        type: string
      list_position:
        type: integer
      name:
        type: string
    type: object
  dto.NewTaskInfo:
    properties:
      description:
        type: string
      end:
        type: string
      list_id:
        type: integer
      list_position:
        type: integer
      name:
        type: string
      start:
        type: string
    type: object
...
  dto.TaskID:
    properties:
      id:
        type: integer
    type: object
...
  dto.UpdatedListInfo:
    properties:
      description:
        type: string
      id:
        type: integer
      list_position:
        type: string
      name:
        type: string
    type: object
  dto.UpdatedTaskInfo:
    properties:
      description:
        type: string
      end:
        type: string
      id:
        type: integer
      list_position:
        type: string
      name:
        type: string
      start:
        type: string
    type: object
...
  entities.List:
    properties:
      board_id:
        type: integer
      description:
        type: string
      id:
        type: integer
      list_position:
        type: integer
      name:
        type: string
      tasks:
        items:
          $ref: '#/definitions/entities.Task'
        type: array
    type: object
  entities.Task:
    properties:
      date_created:
        type: string
      description:
        type: string
      end:
        type: string
      id:
        type: integer
      list_id:
        type: integer
      list_position:
        type: integer
      name:
        type: string
      start:
        type: string
      users:
        items:
          $ref: '#/definitions/entities.User'
        type: array
    type: object
  entities.User:
    properties:
      avatar_url:
        type: string
      description:
        type: string
      email:
        type: string
      name:
        type: string
      password_hash:
        type: string
      surname:
        type: string
      user_id:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: kdanil01@mail.ru
    name: Капитанов Даниил
    url: https://vk.com/poophead27
  description: Лучшее и единственно приложение, имитирующее Trello.
  license:
    name: None
    url: None
  title: LA TABULA API
  version: "2.0"
paths:
...
  /list/create/:
    post:
      consumes:
      - application/json
      description: Создать список
      parameters:
      - description: данные нового списка
        in: body
        name: newListInfo
        required: true
        schema:
          $ref: '#/definitions/dto.NewListInfo'
      produces:
      - application/json
      responses:
        "200":
          description: объект списка
          schema:
            $ref: '#/definitions/doc_structs.ListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      summary: Создать список
      tags:
      - lists
  /list/delete/:
    delete:
      consumes:
      - application/json
      description: Удалить список
      parameters:
      - description: id списка
        in: body
        name: listID
        required: true
        schema:
          $ref: '#/definitions/dto.ListID'
      produces:
      - application/json
      responses:
        "204":
          description: no content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      summary: Удалить список
      tags:
      - lists
  /list/update/:
    post:
      consumes:
      - application/json
      description: Обновить список
      parameters:
      - description: обновленные данные списка
        in: body
        name: listInfo
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatedListInfo'
      produces:
      - application/json
      responses:
        "204":
          description: no content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      summary: Обновить список
      tags:
      - lists
  /task/:
    post:
      consumes:
      - application/json
      description: Получить задание
      parameters:
      - description: id задания
        in: body
        name: taskID
        required: true
        schema:
          $ref: '#/definitions/dto.TaskID'
      produces:
      - application/json
      responses:
        "200":
          description: объект задания
          schema:
            $ref: '#/definitions/doc_structs.TaskResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      summary: Получить задание
      tags:
      - tasks
  /task/create/:
    post:
      consumes:
      - application/json
      description: Создать задание
      parameters:
      - description: данные нового задания
        in: body
        name: newTaskInfo
        required: true
        schema:
          $ref: '#/definitions/dto.NewTaskInfo'
      produces:
      - application/json
      responses:
        "200":
          description: объект задания
          schema:
            $ref: '#/definitions/doc_structs.TaskResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      summary: Создать задание
      tags:
      - tasks
  /task/delete/:
    delete:
      consumes:
      - application/json
      description: Удалить задание
      parameters:
      - description: id задания
        in: body
        name: taskID
        required: true
        schema:
          $ref: '#/definitions/dto.TaskID'
      produces:
      - application/json
      responses:
        "204":
          description: no content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      summary: Удалить задание
      tags:
      - tasks
  /task/update/:
    post:
      consumes:
      - application/json
      description: Обновить задание
      parameters:
      - description: обновленные данные задания
        in: body
        name: taskInfo
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatedTaskInfo'
      produces:
      - application/json
      responses:
        "204":
          description: no content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      summary: Обновить задание
      tags:
      - tasks
...
swagger: "2.0"